+++ changelog.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Added
+
+- [Add support for custom time duration calculations](https://github.com/ballerina-platform/ballerina-library/issues/6840)
+
 ## [2.5.0] - 2024-09-12
 ### Fixed
 - [When converting a `time:Civil` with time-zone information to a string using `time:civilToString` API error is thrown](https://github.com/ballerina-platform/ballerina-library/issues/6986)